scape_project/requirements.txt
django==4.1.1
django_filter==22.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
djangorestframework-gis==1.0
djoser==2.0.5
drf-yasg==1.21.4
psycopg2_binary==2.9.5
JSONField